feat: diff based L1 attributes transaction
Adds an experimental spec for a diff based L1 Attributes transaction.
This helps to reduce the historical growth of the chain by making
the L1 attributes transaction only push values that have changed.

This spec is placed in an `experimental` directory temporarily
until we figure out better ways to manage possible changes to the
protocol. By creating the spec now, we can more easily schedule
it for a future hardfork.